% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emarinade-package.R
\docType{package}
\name{emarinade}
\alias{emarinade-package}
\alias{emarinade}
\title{emarinade: Explore Marine Non-Indigenous Species Data}
\description{
The \strong{emarinade} package provides:
}
\details{
\itemize{
\item Curated datasets: \code{taxo_tbl}, \code{inv_tbl}, \code{origin_tbl}, \code{meow}
\item A Shiny app for species search, taxonomy exploration, and mapping
\item Functions for working with invasion and native range data
}
}
\section{Datasets}{

\itemize{
\item \code{taxo_tbl}: Taxonomic information
\item \code{inv_tbl}: Invasion records
\item \code{origin_tbl}: Native range records
\item \code{meow}: Marine ecoregions polygons
}
}

\author{
\strong{Maintainer}: Clément Violet \email{clement.violet@ifremer.fr} (\href{https://orcid.org/0000-0001-5000-0004}{ORCID})

Authors:
\itemize{
  \item Mathieu Chevalier (\href{https://orcid.org/0000-0001-5000-0001}{ORCID})
  \item Amelia Curd (\href{https://orcid.org/0000-0001-5000-0001}{ORCID})
  \item Frédérique Viard (\href{https://orcid.org/0000-0001-5000-0001}{ORCID})
}

Other contributors:
\itemize{
  \item Mathisse Meyneng (\href{https://orcid.org/0000-0001-5000-0004}{ORCID}) (Logo designer) [contributor]
}

}
\keyword{package}
