% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow.R
\docType{data}
\name{meow}
\alias{meow}
\title{Marine Ecoregion Of the World}
\format{
\subsection{\code{meow}}{

A \code{tibble} data frame with 232 rows and 9 columns:
\describe{
\item{ECO_CODE}{Code of the Province.}
\item{ECOREGION}{Ecoregion name.}
\item{PROV_CODE}{Code of the Province.}
\item{PROVINCE}{Province name.}
\item{RLM_CODE}{Code of the Realm.}
\item{REALM}{Realm Name.}
\item{ECO_CODE_X}{Code of the Ecoregion in a short format.}
\item{Lat_Zone}{Latitudinal zone where the Ecoregion is located.}
\item{WKT}{Well-known text describing the Ecoregion polygon. The CRS is WGS84.}
}
}
}
\source{
Spalding, Mark D., Helen E. Fox, Gerald R. Allen, Nick Davidson, Zach A. Ferdaña, Max Finlayson, Benjamin S. Halpern, Miguel A. Jorge, Al Lombana, Sara A. Lourie, Kirsten D. Martin, Edmund McManus, Jennifer Molnar, Cheri A. Recchia, and James Robertson. 2007. “Marine Ecoregions of the World: A Bioregionalization of Coastal and Shelf Areas.” BioScience 57(7):573–83. doi:\href{https://doi.org/10.1641/B570707}{10.1641/B570707}.
}
\usage{
meow
}
\description{
A \verb{sf object} containing the Marine Ecoregions of the World described by @Spalding_2007, a nested system of realms, provinces, and ecoregions This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
