% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_id_tbl.R
\docType{data}
\name{taxo_id_tbl}
\alias{taxo_id_tbl}
\title{Taxonomic Identifiers Information data}
\format{
\subsection{\code{taxo_id_tbl}}{

A \code{tibble} data frame with 3,511 rows and 5 columns:
\describe{
\item{identifierID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{taxonID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{title}{An optional display label for the URL that the publisher may prefer be displayed with the identifier or link.}
\item{identifier}{Other known identifier used for the same taxon. Can be a URL pointing to a webpage, an xml or rdf document, a DOI, UUID or any other identifer.}
\item{subject}{Keywords qualifying the identifier}
}
}
}
\source{
E-MARINADE package
}
\usage{
taxo_id_tbl
}
\description{
A \code{tibble} containing the information on the taxonomic identifiers of taxa present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
