% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_tbl.R
\docType{data}
\name{taxo_tbl}
\alias{taxo_tbl}
\title{Taxonomic Information data}
\format{
\subsection{\code{taxo_tbl}}{

A \code{tibble} data frame with  rows and 18 columns:
\describe{
\item{taxonID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{Kingdom, Phylum, Class, Order, Family, Genus}{The full scientific name in which the taxon is classified}
\item{acceeptedNameUsage}{The full name, with authorship and date information if known, of the currently valid (zoological) or accepted (botanical) taxon.}
}
}
}
\source{
E-MARINADE package
}
\usage{
taxo_tbl
}
\description{
A \code{tibble} containing the information on the taxonomy of taxa present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
