% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy Information data}
\format{
\subsection{\code{taxonomy}}{

A \code{tibble} data frame with 2,016 rows and 8 columns:
\describe{
\item{taxonID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{Kingdom, Phylum, Class, Order, Family, Genus}{The full scientific name in which the taxon is classified}
\item{acceeptedNameUsage}{The full name, with authorship and date information if known, of the currently valid (zoological) or accepted (botanical) taxon.}
}
}
}
\source{
ANIS-E package
}
\usage{
taxonomy
}
\description{
A \code{tibble} containing the information on the taxonomy of taxa present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
