% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_identifiers.R
\docType{data}
\name{taxonomy_identifiers}
\alias{taxonomy_identifiers}
\title{Taxonomic Identifiers Information data}
\format{
\subsection{\code{taxonomy_identifiers}}{

A \code{tibble} data frame with 5,893 rows and 5 columns:
\describe{
\item{identifierID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{taxonID}{An identifier for the set of taxon information. Identifier specific to the data set.}
\item{title}{An optional display label for the URL that the publisher may prefer be displayed with the identifier or link.}
\item{identifier}{Other known identifier used for the same taxon. Can be a URL pointing to a webpage, an xml or rdf document, a DOI, UUID or any other identifer.}
\item{subject}{Keywords qualifying the identifier}
}
}
}
\source{
ANIS-E package
}
\usage{
taxonomy_identifiers
}
\description{
A \code{tibble} containing the information on the taxonomic identifiers of taxa present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
