% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anise-package.R
\docType{package}
\name{anise}
\alias{anise}
\alias{anise-package}
\title{anise: Explore Marine Non-Indigenous Species Data}
\description{
The \code{anise} package provides:
\itemize{
\item Curated datasets: \code{taxo_tbl}, \code{inv_tbl}, \code{origin_tbl}, \code{meow}
\item A Shiny app for species search, taxonomy exploration, and mapping
\item Functions for working with invasion and native range data
}
}
\section{Datasets}{

\itemize{
\item \code{taxo_tbl}: Taxonomic information
\item \code{inv_tbl}: Invasion records
\item \code{origin_tbl}: Native range records
\item \code{meow}: Marine ecoregions polygons
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://clementviolet.github.io/ANIS-E/}
  \item \url{https://github.com/clementviolet/ANIS-E}
  \item Report bugs at \url{https://github.com/clementviolet/ANIS-E/issues}
}

}
\author{
\strong{Maintainer}: Clément Violet \email{clement.violet@umontpellier.fr} (\href{https://orcid.org/0000-0001-6217-5891}{ORCID})

Authors:
\itemize{
  \item Clément Violet \email{clement.violet@umontpellier.fr} (\href{https://orcid.org/0000-0001-6217-5891}{ORCID})
  \item Mathieu Chevalier (\href{https://orcid.org/0000-0002-1170-5343}{ORCID})
  \item Amelia Curd (\href{https://orcid.org/0000-0003-3260-7192}{ORCID})
  \item Frédérique Viard (\href{https://orcid.org/0000-0001-5603-9527}{ORCID})
}

Other contributors:
\itemize{
  \item Mathisse Meyneng (\href{https://orcid.org/0009-0000-7442-4613}{ORCID}) (Other: Logo designer) [contributor]
}

}
\keyword{internal}
\keyword{package}
