% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anise.R
\name{data_anise}
\alias{data_anise}
\title{Loading all anise datasets}
\usage{
data_anise(envir = .GlobalEnv)
}
\arguments{
\item{envir}{The environment where the datasets will be loaded. Defaults to the global environment.}
}
\value{
Load the \code{taxonomy}, \code{taxonomy_identifiers},\code{introduction}, \code{origin} and \code{meow} datasets.
}
\description{
Loading all anise datasets
}
\examples{
\dontrun{
  data_anise()
}
}
\seealso{
Dataset documentation: \link{taxonomy}, \link{taxonomy_identifiers}, \link{introduction}, \link{origin}, \link{meow}.
}
