% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emarinade-package.R
\docType{package}
\name{emarinade}
\alias{emarinade}
\alias{emarinade-package}
\title{emarinade: Explore Marine Non-Indigenous Species Data}
\description{
The \strong{emarinade} package provides:
\itemize{
\item Curated datasets: \code{taxo_tbl}, \code{inv_tbl}, \code{origin_tbl}, \code{meow}
\item A Shiny app for species search, taxonomy exploration, and mapping
\item Functions for working with invasion and native range data
}
}
\section{Datasets}{

\itemize{
\item \code{taxo_tbl}: Taxonomic information
\item \code{inv_tbl}: Invasion records
\item \code{origin_tbl}: Native range records
\item \code{meow}: Marine ecoregions polygons
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://clementviolet.github.io/E-MARINADE/}
}

}
\author{
\strong{Maintainer}: Clément Violet \email{clement.violet@ifremer.fr} (\href{https://orcid.org/0000-0001-6217-5891}{ORCID})

Authors:
\itemize{
  \item Mathieu Chevalier (\href{https://orcid.org/0000-0002-1170-5343}{ORCID})
  \item Amelia Curd (\href{https://orcid.org/0000-0003-3260-7192}{ORCID})
  \item Frédérique Viard (\href{https://orcid.org/0000-0001-5603-9527}{ORCID})
}

Other contributors:
\itemize{
  \item Mathisse Meyneng (\href{https://orcid.org/0009-0000-7442-4613}{ORCID}) (Logo designer) [contributor]
}

}
\keyword{internal}
\keyword{package}
