% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_tbl.R
\docType{data}
\name{inv_tbl}
\alias{inv_tbl}
\title{Invasion Event data}
\format{
\subsection{\code{inv_tbl}}{

A \code{tibble} data frame with 4,834 rows and 18 columns:
\describe{
\item{InvID}{Unique Identifier of an introduction event	numeric}
\item{SpeciesID}{Identifier linking a species to a valid taxonomic}
\item{Year}{Year of the first repported occurence}
\item{Country}{Country where the invasion event occurred}
\item{Status}{Origin status of the species (i.e. Non-indigenous or Cryptogenic)}
\item{Release,Escape, Transport_contaminant, Transport_stowaway, Corridor, Unaided, Unknown}{Pathways and vectors of introduction}
\item{MSFD_region}{Name of the  MSFD region. See \href{https://www.eea.europa.eu/en/datahub/datahubitem-view/7144675c-5c84-456f-92e0-8f832239d880}{here} for details}
\item{MSFD_subregion}{Name of the  MSFD region. See \href{https://www.eea.europa.eu/en/datahub/datahubitem-view/7144675c-5c84-456f-92e0-8f832239d880}{here} for details}
\item{Ecoregion_Code}{Code of the Ecoregion (sensus Marine Ecoregion of the World according to \href{https://doi.org/10.1641/B570707}{Spalding et al. 2007}) of the corresponding MSFD sub-region.}
\item{Source}{Reference indicating that the species is a cryptogenic or non-indigenous species in the area}
\item{DB}{Reference (e.g., database, review, or synthesis publication) compiling or citing the Source that indicates the species is cryptogenic or non-indigenous in the area}
}
}
}
\source{
E-MARINADE package
}
\usage{
inv_tbl
}
\description{
A \code{tibble} containing the introduction events present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
