% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_tbl.R
\docType{data}
\name{taxo_tbl}
\alias{taxo_tbl}
\title{Taxonomic Information data}
\format{
\subsection{\code{inv_tbl}}{

A \code{tibble} data frame with 4,834 rows and 18 columns:
\describe{
\item{SpeciesID}{Identifier linking a species to a valid taxonomic}
\item{AphiaID}{Identifier linking a species to a valid taxonomic}
\item{tsnID}{ITIS identifier.}
\item{boldID}{Barcode of Life identifier.}
\item{eolID}{Encyclopedia of Life identifier.}
\item{fishbaseID}{Fishbase identifier.}
\item{algaebaseID}{Algaebase identifier.}
\item{iucnID}{IUCN Red List identifier.}
\item{ncbiID}{NCBI identifier.}
\item{Kingdom, Phylum, Class, Order, Family, Genus, Species}{Taxonomic rank.}#'
}
}
}
\source{
E-MARINADE package
}
\usage{
taxo_tbl
}
\description{
A \code{tibble} containing the information on the taxonomy of taxa present in the Shiny app. This dataset can be useful when working with the database programmatically.
}
\keyword{datasets}
